+++ fidget-raster/src/voxel.rs
@@ -1,8 +1,8 @@
 //! 3D bitmap rendering / rasterization
 use super::RenderHandle;
 use crate::{
-    Image as GenericImage, RenderConfig as RenderConfigLike, RenderWorker,
-    Tile, TileSizesRef,
+    Image as GenericImage, RenderConfig as RenderConfigLike, RenderError,
+    RenderWorker, Tile, TileSizesRef,
 };
 use fidget_core::{
     eval::Function,
@@ -80,8 +80,8 @@ impl RenderConfigLike for RenderConfig<'_> {
 impl RenderConfig<'_> {
     /// Render a shape in 3D using this configuration
     ///
-    /// Returns an [`Image`] of pixel data, or `None` if rendering was
-    /// cancelled.
+    /// Returns [`Ok(Some(Image))`](Image) of pixel data on success, `Ok(None)`
+    /// if the render was cancelled, or an error.
     ///
     /// In the resulting image, saturated pixels (i.e. pixels in the image which
     /// are fully occupied up to the camera) are represented with `depth =
@@ -89,7 +89,7 @@ impl RenderConfig<'_> {
     pub fn run<F: Function + RenderHints>(
         &self,
         shape: Shape<F>,
-    ) -> Option<Image> {
+    ) -> Result<Option<Image>, RenderError> {
         self.run_with_vars::<F>(shape, &ShapeVars::new())
     }
 
@@ -98,7 +98,7 @@ impl RenderConfig<'_> {
         &self,
         shape: Shape<F>,
         vars: &ShapeVars<f32>,
-    ) -> Option<Image> {
+    ) -> Result<Option<Image>, RenderError> {
         render(shape, vars, self)
     }
 
@@ -488,23 +488,29 @@ impl<F: Function> Worker<'_, F> {
 /// This function is parameterized by shape type, which determines how we
 /// perform evaluation.
 ///
-/// Returns a [`Image`] of pixels, or `None` if rendering was cancelled
-/// (using the [`RenderConfig::cancel`] token)
+/// Returns [`Ok(Some(Image))`](Image) of pixel data on success, `Ok(None)` if
+/// the render was cancelled, or an error.
 pub fn render<F: Function + RenderHints>(
     shape: Shape<F>,
     vars: &ShapeVars<f32>,
     config: &RenderConfig,
-) -> Option<Image> {
+) -> Result<Option<Image>, RenderError> {
+    vars.check(&shape)?;
     let max_size = config.width().max(config.height()) as usize;
     let default_tile_sizes;
+
     let tile_sizes = if let Some(ts) = &config.tile_sizes {
         TileSizesRef::new(ts, max_size)
     } else {
         default_tile_sizes = F::tile_sizes_3d();
         TileSizesRef::new(&default_tile_sizes, max_size)
     };
-    let tiles =
-        super::render_tiles::<F, Worker<F>>(shape, vars, config, tile_sizes)?;
+    let tiles = match super::render_tiles::<F, Worker<F>>(
+        shape, vars, config, tile_sizes,
+    ) {
+        Some(t) => t,
+        None => return Ok(None),
+    };
 
     let width = config.image_size.width() as usize;
     let height = config.image_size.height() as usize;
@@ -534,13 +540,13 @@ pub fn render<F: Function + RenderHints>(
             }
         }
     }
-    Some(image)
+    Ok(Some(image))
 }
 
 #[cfg(test)]
 mod test {
     use super::*;
-    use fidget_core::{Context, vm::VmShape};
+    use fidget_core::{Context, var::Var, vm::VmShape};
 
     /// Make sure we don't crash if there's only a single tile
     #[test]
@@ -553,7 +559,10 @@ mod test {
             image_size: RenderSize::from(128), // very small!
             ..Default::default()
         };
-        let image = cfg.run(shape).unwrap();
+        let image = cfg
+            .run(shape)
+            .expect("rendering should not fail")
+            .expect("rendering should not be cancelled");
         assert_eq!(image.len(), 128 * 128);
     }
 
@@ -569,7 +578,37 @@ mod test {
         };
         let cancel = cfg.cancel.clone();
         cancel.cancel();
-        let out = cfg.run::<_>(shape);
-        assert!(out.is_none());
+        assert!(cfg.run::<_>(shape).unwrap().is_none());
+    }
+
+    #[test]
+    fn missing_var() {
+        let mut ctx = Context::new();
+        let x = ctx.x();
+        let v = ctx.var(Var::new());
+        let s = ctx.sub(x, v).unwrap();
+        let shape = VmShape::new(&ctx, s).unwrap();
+
+        let cfg = RenderConfig {
+            image_size: RenderSize::new(64, 64, 64),
+            ..Default::default()
+        };
+        let Err(out) = cfg.run::<_>(shape.clone()) else {
+            panic!("expected error")
+        };
+        let var = ctx.get_var(v).unwrap();
+        let Var::V(i) = var else {
+            panic!("expected Var::V")
+        };
+        assert_eq!(
+            out,
+            RenderError::MissingVar(fidget_core::shape::MissingVar { var: i })
+        );
+
+        let mut vars = ShapeVars::new();
+        vars.insert(i, 1.0);
+        cfg.run_with_vars::<_>(shape, &vars)
+            .expect("rendering worked")
+            .expect("not cancelled");
     }
 }